Updates to cleanup the CampaignSubPage component.
